requirements.txt
flask
python-aiml
scikit-learn
pyngrok